* main:
  Update README.md
  bump mechanic
  update oid to long in mapstores
  update serializers to use long for oid
  update oid param to long in getOrCreateTable
  update getDatabaseOid to return long in DatabaseQueryManager interface
  update getDatabaseOid to use oid adapter and return a long
  update oid to long in TableInfo
  update oid to long in OrganizationDatabase
  update oid to long in ExternalTable
  switch oid from int to long
  Update conductor-client/src/main/kotlin/com/openlattice/graph/Graph.kt
  Ensure edge deletions are propagated to all data sources.
  Add data integration controller to list of class path scanned base packages.
  Fix data integration controller package name.
  Initialize things for stream serializer.
  remove unused import
  Remove java elasticsearch impl.

+++ conductor-client/src/main/kotlin/com/openlattice/hazelcast/serializers/OrganizationDatabaseStreamSerializer.kt
@@ -23,12 +23,12 @@ class OrganizationDatabaseStreamSerializer : TestableSelfRegisteringStreamSerial
     }
 
     override fun write(out: ObjectDataOutput, `object`: OrganizationDatabase) {
-        out.writeInt(`object`.oid)
+        out.writeLong(`object`.oid)
         out.writeUTF(`object`.name)
     }
 
     override fun read(`in`: ObjectDataInput): OrganizationDatabase {
-        val oid = `in`.readInt()
+        val oid = `in`.readLong()
         val name = `in`.readUTF()
         return OrganizationDatabase(oid, name)
     }
